t6120: test describe and name-rev with deep repos
Depending on the implementation of walks, limitted stack size may lead
to problems (for recursion).

Test name-rev and describe with deep repos and limitted stack size and
mark the former with known failure.

We add these tests (which add gazillions of commits) lasts so as to keep
the runtime of other subtests the same.