log: show ^ for boundary
Boundary commits are commits which are discarded by the revision walk,
i.e. they are no "positive part" of the specified revision set but show
nevertheless (because of `--boundary`).

So far, `log` shows them with a `-` prefix and `log --graph` with a `o`.
besides the obvious inconsistency, these choices are bad for other
reasons, too:

`-` is typically used in a diff whereas negative revs are denoted by
`^`.

`o` reminds of "origin/origo" at least as much as of "omitted";
visually, it's not clear that it "should not be present".

So, change both `log` and `log --graph` to use the same symbol `^` which
the revision parser uses for negative revs, too.