log --graph: denote root commits by o
So far, root commits are not distinguished in log output at all. As a
consequence, you cannot tell apart one linearly connected history from
two disconnect linear histories at all.

Change this so that root commits (i.e. those with 0 parents) are denoted
by `o` for "origo/origin" (or, if you like, a small zero). As a first
step, we do this for `log --graph` only.