status: show in-progress info for porcelain v2
Porcelain v2 format is intended to be extendable. So we add in-progress
information unconditionally to the branch information (in case that is
shown at all).

Note that - in contrast to the short status info - we add all available
information in case multiple operations are in progress, and we do so by
outputting one line per operation.